refactor(dashboard): extract plugin field schemas from dslSchemas
dslSchemas.ts's getPluginFieldSchema had grown into an 800-line file
with a single ~340-line function covering every tool_selection and
related plugin type, tripping this repo's file-length and
function-length structure caps. Extracted it into
dashboard/frontend/src/lib/dslPluginSchemas.ts as a short delegating
switch plus one small helper per plugin type
(getMemoryPluginFieldSchema, getToolsPluginFieldSchema, etc.),
matching this directory's narrow-module convention
(dashboard/frontend/src/AGENTS.md: keep page-specific/feature-local
logic out of shared modules, in sibling support modules instead).
dslSchemas.ts re-exports getPluginFieldSchema so existing importers
(dslMutations.ts and its own downstream consumers) need no changes.
Pure move, no behavior change -- every case's field array is
unchanged, including the sticky tool-set selection block under
tool_selection.

Verified with the real toolchain, not just structure_check.py: tsc
--noEmit clean, eslint clean on both files, and make dashboard-check
end to end (catalog check, lint, type-check, the full frontend unit
suite -- 726 tests across 171 files, including the tests that
exercise getPluginFieldSchema directly -- and dashboard/backend's Go
suite) all pass unchanged.
